Small Memory Leak Fixes & Fix VS Project Paths (#18)
* Fix Emulator VS Project Include Paths

The previous paths were hardcoded to local directories which prevented
the solution from compiling properly on other machines.

This change updates the paths to use VS path macros instead which makes
the include paths relative to the project directory.

* Fix Small Memory Leak in RV32 Emulator

The SDecodedBlock structures allocated as part of m_decodedBlocks were
never being deleted when the map was cleared (during reset and fence
instructions).

This change adds logic to delete the structures whenever the map is
cleared which resolves the leak.

* Add Various Missing Deletes

Several places in the code were missing delete logic which were
identified while investigating an unrelated memory leak.

* Remove Incorrect CSR Mem Reset

There's a reset for a CSRMem at index 2 but the bus only allocates 2
CSRMem pointers so the third one (at index 2) is invalid.

Thanks to Chris--B for finding this with UBSan.

---------

Co-authored-by: Engin Cilasun <ecilasun@me.com>